Tests: Forms: Bulk & Quick Edit: Refactor Tests
