--- tests/EndToEnd/forms/post-types/PageBulkQuickEditFormCest.php
@@ -1,324 +1,0 @@
-<?php
-
-namespace Tests\EndToEnd;
-
-use Tests\Support\EndToEndTester;
-
-/**
- * Tests for Kit Forms on WordPress Pages when using the Bulk and Quick Edit functionality.
- *
- * @since   1.9.6
- */
-class PageBulkQuickEditFormCest
-{
-	/**
-	 * Run common actions before running the test functions in this class.
-	 *
-	 * @since   1.9.6
-	 *
-	 * @param   EndToEndTester $I  Tester.
-	 */
-	public function _before(EndToEndTester $I)
-	{
-		// Activate Kit plugin.
-		$I->activateKitPlugin($I);
-	}
-
-	/**
-	 * Test that the Default Form for Pages displays when the Default option is chosen via
-	 * WordPress' Quick Edit functionality.
-	 *
-	 * @since   1.9.8.0
-	 *
-	 * @param   EndToEndTester $I  Tester.
-	 */
-	public function testQuickEditUsingDefaultForm(EndToEndTester $I)
-	{
-		// Setup Kit plugin.
-		$I->setupKitPlugin($I);
-		$I->setupKitPluginResources($I);
-
-		// Programmatically create a Page.
-		$pageID = $I->havePostInDatabase(
-			[
-				'post_type'  => 'page',
-				'post_title' => 'Kit: Page: Form: Default: Quick Edit',
-			]
-		);
-
-		// Quick Edit the Page in the Pages WP_List_Table.
-		$I->quickEdit(
-			$I,
-			postType: 'page',
-			postID: $pageID,
-			configuration: [
-				'form' => [ 'select', 'Default' ],
-			]
-		);
-
-		// Load the Page on the frontend site.
-		$I->amOnPage('/?p=' . $pageID);
-
-		// Check that no PHP warnings or notices were output.
-		$I->checkNoWarningsAndNoticesOnScreen($I);
-
-		// Confirm that one Kit Form is output in the DOM.
-		// This confirms that there is only one script on the page for this form, which renders the form.
-		$I->seeFormOutput($I, $_ENV['CONVERTKIT_API_FORM_ID']);
-	}
-
-	/**
-	 * Test that the defined form displays when chosen via
-	 * WordPress' Quick Edit functionality.
-	 *
-	 * @since   1.9.8.0
-	 *
-	 * @param   EndToEndTester $I  Tester.
-	 */
-	public function testQuickEditUsingDefinedForm(EndToEndTester $I)
-	{
-		// Setup Kit plugin.
-		$I->setupKitPlugin($I);
-		$I->setupKitPluginResources($I);
-
-		// Programmatically create a Page.
-		$pageID = $I->havePostInDatabase(
-			[
-				'post_type'  => 'page',
-				'post_title' => 'Kit: Page: Form: ' . $_ENV['CONVERTKIT_API_FORM_NAME'] . ': Quick Edit',
-			]
-		);
-
-		// Quick Edit the Page in the Pages WP_List_Table.
-		$I->quickEdit(
-			$I,
-			postType: 'page',
-			postID: $pageID,
-			configuration: [
-				'form' => [ 'select', $_ENV['CONVERTKIT_API_FORM_NAME'] ],
-			]
-		);
-
-		// Load the Page on the frontend site.
-		$I->amOnPage('/?p=' . $pageID);
-
-		// Check that no PHP warnings or notices were output.
-		$I->checkNoWarningsAndNoticesOnScreen($I);
-
-		// Confirm that one Kit Form is output in the DOM.
-		// This confirms that there is only one script on the page for this form, which renders the form.
-		$I->seeFormOutput($I, $_ENV['CONVERTKIT_API_FORM_ID']);
-	}
-
-	/**
-	 * Test that the Default Form for Pages displays when the Default option is chosen via
-	 * WordPress' Bulk Edit functionality.
-	 *
-	 * @since   1.9.8.0
-	 *
-	 * @param   EndToEndTester $I  Tester.
-	 */
-	public function testBulkEditUsingDefaultForm(EndToEndTester $I)
-	{
-		// Setup Kit plugin.
-		$I->setupKitPlugin($I);
-		$I->setupKitPluginResources($I);
-
-		// Programmatically create two Pages.
-		$pageIDs = array(
-			$I->havePostInDatabase(
-				[
-					'post_type'  => 'page',
-					'post_title' => 'Kit: Page: Form: Default: Bulk Edit #1',
-				]
-			),
-			$I->havePostInDatabase(
-				[
-					'post_type'  => 'page',
-					'post_title' => 'Kit: Page: Form: Default: Bulk Edit #2',
-				]
-			),
-		);
-
-		// Bulk Edit the Pages in the Pages WP_List_Table.
-		$I->bulkEdit(
-			$I,
-			postType: 'page',
-			postIDs: $pageIDs,
-			configuration: [
-				'form' => [ 'select', 'Default' ],
-			]
-		);
-
-		// Iterate through Pages to run frontend tests.
-		foreach ($pageIDs as $pageID) {
-			// Load Page on the frontend site.
-			$I->amOnPage('/?p=' . $pageID);
-
-			// Check that no PHP warnings or notices were output.
-			$I->checkNoWarningsAndNoticesOnScreen($I);
-
-			// Confirm that one Kit Form is output in the DOM.
-			// This confirms that there is only one script on the page for this form, which renders the form.
-			$I->seeFormOutput($I, $_ENV['CONVERTKIT_API_FORM_ID']);
-		}
-	}
-
-	/**
-	 * Test that the defined form displays when chosen via
-	 * WordPress' Bulk Edit functionality.
-	 *
-	 * @since   1.9.8.0
-	 *
-	 * @param   EndToEndTester $I  Tester.
-	 */
-	public function testBulkEditUsingDefinedForm(EndToEndTester $I)
-	{
-		// Setup Kit plugin.
-		$I->setupKitPlugin($I);
-		$I->setupKitPluginResources($I);
-
-		// Programmatically create two Pages.
-		$pageIDs = array(
-			$I->havePostInDatabase(
-				[
-					'post_type'  => 'page',
-					'post_title' => 'Kit: Page: Form: ' . $_ENV['CONVERTKIT_API_FORM_NAME'] . ': Bulk Edit #1',
-				]
-			),
-			$I->havePostInDatabase(
-				[
-					'post_type'  => 'page',
-					'post_title' => 'Kit: Page: Form: ' . $_ENV['CONVERTKIT_API_FORM_NAME'] . ': Bulk Edit #2',
-				]
-			),
-		);
-
-		// Bulk Edit the Pages in the Pages WP_List_Table.
-		$I->bulkEdit(
-			$I,
-			postType: 'page',
-			postIDs: $pageIDs,
-			configuration: [
-				'form' => [ 'select', $_ENV['CONVERTKIT_API_FORM_NAME'] ],
-			]
-		);
-
-		// Iterate through Pages to run frontend tests.
-		foreach ($pageIDs as $pageID) {
-			// Load Page on the frontend site.
-			$I->amOnPage('/?p=' . $pageID);
-
-			// Check that no PHP warnings or notices were output.
-			$I->checkNoWarningsAndNoticesOnScreen($I);
-
-			// Confirm that one Kit Form is output in the DOM.
-			// This confirms that there is only one script on the page for this form, which renders the form.
-			$I->seeFormOutput($I, $_ENV['CONVERTKIT_API_FORM_ID']);
-		}
-	}
-
-	/**
-	 * Test that the existing settings are honored and not changed
-	 * when the Bulk Edit options are set to 'No Change'.
-	 *
-	 * @since   1.9.8.0
-	 *
-	 * @param   EndToEndTester $I  Tester.
-	 */
-	public function testBulkEditWithNoChanges(EndToEndTester $I)
-	{
-		// Setup Kit plugin.
-		$I->setupKitPlugin($I);
-		$I->setupKitPluginResources($I);
-
-		// Programmatically create two Pages with a defined form.
-		$pageIDs = array(
-			$I->havePostInDatabase(
-				[
-					'post_type'  => 'page',
-					'post_title' => 'Kit: Page: Form: ' . $_ENV['CONVERTKIT_API_FORM_NAME'] . ': Bulk Edit with No Change #1',
-					'meta_input' => [
-						'_wp_convertkit_post_meta' => [
-							'form'         => $_ENV['CONVERTKIT_API_FORM_ID'],
-							'landing_page' => '',
-							'tag'          => '',
-						],
-					],
-				]
-			),
-			$I->havePostInDatabase(
-				[
-					'post_type'  => 'page',
-					'post_title' => 'Kit: Page: Form: ' . $_ENV['CONVERTKIT_API_FORM_NAME'] . ': Bulk Edit with No Change #2',
-					'meta_input' => [
-						'_wp_convertkit_post_meta' => [
-							'form'         => $_ENV['CONVERTKIT_API_FORM_ID'],
-							'landing_page' => '',
-							'tag'          => '',
-						],
-					],
-				]
-			),
-		);
-
-		// Bulk Edit the Pages in the Pages WP_List_Table.
-		$I->bulkEdit(
-			$I,
-			postType: 'page',
-			postIDs: $pageIDs,
-			configuration: [
-				'form' => [ 'select', '— No Change —' ],
-			]
-		);
-
-		// Iterate through Pages to run frontend tests.
-		foreach ($pageIDs as $pageID) {
-			// Load Page on the frontend site.
-			$I->amOnPage('/?p=' . $pageID);
-
-			// Check that no PHP warnings or notices were output.
-			$I->checkNoWarningsAndNoticesOnScreen($I);
-
-			// Confirm that one Kit Form is output in the DOM.
-			// This confirms that there is only one script on the page for this form, which renders the form.
-			$I->seeFormOutput($I, $_ENV['CONVERTKIT_API_FORM_ID']);
-		}
-	}
-
-	/**
-	 * Test that the Bulk Edit fields do not display when a search on a WP_List_Table
-	 * returns no results.
-	 *
-	 * @since   1.9.8.1
-	 *
-	 * @param   EndToEndTester $I  Tester.
-	 */
-	public function testBulkEditFieldsHiddenWhenNoPagesFound(EndToEndTester $I)
-	{
-		// Setup Kit plugin.
-		$I->setupKitPlugin($I);
-		$I->setupKitPluginResources($I);
-
-		// Emulate the user searching for Pages with a query string that yields no results.
-		$I->amOnAdminPage('edit.php?post_type=page&s=nothing');
-
-		// Confirm that the Bulk Edit fields do not display.
-		$I->dontSeeElement('#convertkit-bulk-edit');
-	}
-
-	/**
-	 * Deactivate and reset Plugin(s) after each test, if the test passes.
-	 * We don't use _after, as this would provide a screenshot of the Plugin
-	 * deactivation and not the true test error.
-	 *
-	 * @since   1.9.6.7
-	 *
-	 * @param   EndToEndTester $I  Tester.
-	 */
-	public function _passed(EndToEndTester $I)
-	{
-		$I->deactivateKitPlugin($I);
-		$I->resetKitPlugin($I);
-	}
-}
